refactor: reorganize components into semantic directory structure
