+++ src/components/display/FloatingParticle.tsx
