+++ src/components/display/ScrollIndicator.tsx
