+++ src/components/display/SectionHeading.tsx
