+++ src/components/display/StatusCircle.tsx
