+++ src/components/display/VersionBadge.tsx
