+++ src/components/display/index.ts
@@ -1,0 +1,9 @@
+export { ProjectCard } from './ProjectCard';
+export { StatusBadge } from './StatusBadge';
+export { StatusCircle } from './StatusCircle';
+export { VersionBadge } from './VersionBadge';
+export { PaginationDots } from './PaginationDots';
+export { SectionHeading } from './SectionHeading';
+export { SkillPill } from './SkillPill';
+export { ScrollIndicator } from './ScrollIndicator';
+export { default as FloatingParticle } from './FloatingParticle';
