+++ src/components/interactive/index.ts
@@ -1,0 +1,2 @@
+export { ThemeButton } from './ThemeButton';
+export { SocialIcons } from './SocialIcons';
