+++ src/components/logos/index.ts
@@ -1,0 +1,4 @@
+export { default as GoogleLogo } from './GoogleLogo';
+export { default as MetaLogo } from './MetaLogo';
+export { default as OpenAILogo } from './OpenAILogo';
+export { default as PSULogo } from './PSULogo';
