+++ src/sections/AboutSection.tsx
@@ -1,7 +1,7 @@
 // src/sections/AboutSection.tsx
 
 // imports
-import { SkillPill } from '~/components/ui/SkillPill';
+import { SkillPill } from '~/components/display/SkillPill';
 import { aboutContent } from '~/data/content/about';
 import { skillCategories } from '~/data/structured/skills';
