+++ src/sections/HeroSection.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState, useEffect } from 'react';
 import Image from 'next/image';
-import FloatingParticle from '~/components/ui/FloatingParticle';
+import FloatingParticle from '~/components/display/FloatingParticle';
 import useTypingAnimation from '~/hooks/useTypingAnimation';
 import type { HeroSectionProps } from '~/types';
